修改帖子的接口

+++ src/main/java/org/dgut/community/service/user/impl/UserServiceImpl.java
@@ -73,7 +73,9 @@ public class UserServiceImpl implements IUser {
 //                user.setUserName(newUser.getUserName());
 //            }
             if (newUser.getUserHeadImg() != ""){
-                Util.deleteFile(user.getUserHeadImg());
+                if (user.getUserHeadImg() != null){
+                    Util.deleteFile(user.getUserHeadImg());
+                }
                 user.setUserHeadImg(Util.uploadBase64Image(user.getUserName(), newUser.getUserHeadImg()));
             }
 //            user.setUserName(newUser.getUserName());
